# Performance: Optimize regex processing with compiled RegexSet for batch operations